reorganized file structure so that course commands have their own folder
