+++ src/interaction/command/courseManagement/Command.RemoveCourse.ts
@@ -1,6 +1,6 @@
 import { CommandInteraction, Client, ApplicationCommandOptionType, CategoryChannel } from 'discord.js';
-import { Course, courseModel } from '../../models/Course';
-import { Command } from './interface.Command';
+import { Course, courseModel } from '../../../models/Course';
+import { Command } from '../interface.Command';
 
 
 export const removeCourse: Command = {
