+++ src/interaction/command/data.CommandList.ts
@@ -2,10 +2,10 @@
 import { Command } from "./interface.Command";
 
 // Imported commands
-import { addCourse } from "./Command.addcourse";
+import { addCourse } from "./courseManagement/Command.addcourse";
 import { DiscussionMenu } from "./menu/Command.DiscussionMenu";
 import { testMenu } from "./menu/Command.TestMenu";
-import { removeCourse } from "./Command.RemoveCourse";
+import { removeCourse } from "./courseManagement/Command.RemoveCourse";
 
 /**
  * @constant List of all the available commands
